requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
greenlet==1.1.2
idna==3.3
oauthlib==3.1.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
SQLAlchemy==1.4.29
SQLAlchemy-Utils==0.38.2
tweepy==4.4.0
urllib3==1.26.8
